Move common Schedule models to common/ dir
